Change signature of bfd crc functions
The crc calculated is 32 bits.  Replace uses of unsigned long with
uint32_t.  Also use bfd_byte* for buffers.

bfd/
	* opncls.c (bfd_calc_gnu_debuglink_crc32): Use stdint types.
	(bfd_get_debug_link_info_1, bfd_get_debug_link_info): Likewise.
	(separate_debug_file_exists, bfd_follow_gnu_debuglink): Likewise.
	(bfd_fill_in_gnu_debuglink_section): Likewise.
	* bfd-in2.h: Regenerate.
gdb/
	* auto-load.c (auto_load_objfile_script): Update type of
	bfd_get_debug_link_info argument.
	* symfile.c (find_separate_debug_file_by_debuglink): Likewise.
	* gdb_bfd.c (get_file_crc): Update type of
	bfd_calc_gnu_debuglink_crc32 argument.

+++ gdb/auto-load.c
@@ -856,7 +856,7 @@ auto_load_objfile_script (struct objfile *objfile,
   struct objfile *parent = objfile->separate_debug_objfile_backlink;
   if (parent != nullptr)
     {
-      unsigned long crc32;
+      uint32_t crc32;
       gdb::unique_xmalloc_ptr<char> debuglink
 	(bfd_get_debug_link_info (parent->obfd.get (), &crc32));
